Create Multiple_Lightbulbs
You still must determine how long the room will be lit during the observation period between start_watching and end_watching. But now we have more than one light bulb. This means that in the light bulb switching array can now also be passed the number of the light bulb, the button of which is being pressed.
Each element of the button clicking array can be either a datetime object (which means the time when the first button was pressed) or a tuple of 2 elements (where the first elements is a datetime object, the time the button was pressed), and the second is the number of the light bulb, the button of which is being pressed.
If the passed array will only consist of datetime elements, then we have only one light bulb and the function should work the same way as in the previous mission of the series.
Input: Three arguments and only the first one is required. The first one is a list of datetime objects (instead of datetime object there can be a tuple of two datetime and int), the second and the third ones are the datetime objects.
Output: A number of seconds as an integer